+++ packages/fe-fpm-writer/test/unit/building-block/prompts/utils/questions.test.ts
@@ -21,8 +21,8 @@ import {
     getViewOrFragmentPathPrompt
 } from '../../../../../src/building-block/prompts/utils/questions';
 import type { ListPromptQuestion, PromptContext } from '../../../../../src/prompts/types';
-import { BuildingBlockType } from '../../../../../src';
 import { bindingContextAbsolute, bindingContextRelative } from '../../../../../src/building-block/types';
+import * as promptHelpers from '../../../../../src/building-block/prompts/utils/prompt-helpers';
 
 const projectFolder = join(__dirname, '../../../sample/building-block/webapp-prompts');
 const capProjectFolder = join(__dirname, '../../../sample/building-block/webapp-prompts-cap');
@@ -43,6 +43,11 @@ jest.mock('@sap-ux/project-access', () => ({
     getCapServiceName: jest.fn().mockResolvedValue('mappedMainServiceName')
 }));
 
+jest.mock('../../../../../src/building-block/prompts/utils/prompt-helpers', () => ({
+    ...jest.requireActual('../../../../../src/building-block/prompts/utils/prompt-helpers'),
+    getEntitySetOptions: jest.fn()
+}));
+
 describe('utils - questions', () => {
     let project: Project;
     let capProject: Project;
@@ -59,6 +64,12 @@ describe('utils - questions', () => {
             appPath: projectFolder,
             project
         };
+    });
+
+    beforeEach(() => {
+        (promptHelpers.getEntitySetOptions as jest.Mock).mockImplementation(
+            jest.requireActual('../../../../../src/building-block/prompts/utils/prompt-helpers').getEntitySetOptions
+        );
     });
 
     test('entityPrompt', async () => {
@@ -117,6 +128,51 @@ describe('utils - questions', () => {
             'to_CntctPersnFuncValueHelp',
             'to_CustomerToBusinessPartner'
         ]);
+    });
+
+    test('getBindingContextTypePrompt uses choices passed in properties and overwrites default choices', async () => {
+        (promptHelpers.getEntitySetOptions as jest.Mock).mockReturnValueOnce([]);
+
+        const bindingContextPrompt = getBindingContextTypePrompt({
+            message: 'bindingContext',
+            choices: async () => {
+                return [
+                    { name: 'Absolute', value: bindingContextAbsolute },
+                    { name: 'Relative', value: bindingContextRelative, disabled: true }
+                ];
+            }
+        });
+
+        const choicesFn = bindingContextPrompt.choices as Choices;
+        expect(choicesFn).toBeDefined();
+
+        const choices = await choicesFn();
+        expect(choices).toEqual([
+            { name: 'Absolute', value: bindingContextAbsolute },
+            { name: 'Relative', value: bindingContextRelative, disabled: true }
+        ]);
+    });
+
+    test('entityPrompt fallback to empty array when no options returned', async () => {
+        const contextWithPageContextEntitySet = {
+            ...context,
+            options: {
+                pageContextEntitySet: 'I_CustomerContactOP'
+            }
+        };
+
+        // return empty array
+        (promptHelpers.getEntitySetOptions as jest.Mock).mockReturnValueOnce([]);
+        const entityPrompt = getEntityPrompt(contextWithPageContextEntitySet, { message: 'entity' });
+        const answers = {
+            buildingBlockData: {
+                metaPath: {
+                    bindingContextType: bindingContextRelative
+                }
+            }
+        };
+        const choices = await (entityPrompt.choices as Choices)(answers);
+        expect(choices).toEqual([]);
     });
 
     test('getTargetPropertiesPrompt returns all properties for absolute binding', async () => {