+++ packages/fiori-annotation-api/src/avt/annotations.ts
@@ -166,6 +166,7 @@ function mergeAnnotation(
         target.value = source.value;
         target.origin = source.origin;
     } else if (source.collection) {
+        target.origin = source.origin;
         target.collection = source.collection;
         target.collectionOrigins = source.collectionOrigins;
     }