+++ packages/generator-adp/src/app/index.ts
@@ -1,6 +1,7 @@
-import { AppWizard, MessageType, Prompts as YeomanUiSteps, type IPrompt } from '@sap-devx/yeoman-ui-types';
+import fs from 'node:fs';
 import { join } from 'node:path';
 import Generator from 'yeoman-generator';
+import { AppWizard, MessageType, Prompts as YeomanUiSteps, type IPrompt } from '@sap-devx/yeoman-ui-types';
 
 import {
     FlexLayer,
@@ -8,16 +9,17 @@ import {
     SystemLookup,
     fetchPublicVersions,
     generate,
+    generateCf,
+    getCfConfig,
     getConfig,
     getConfiguredProvider,
-    isCFEnvironment,
+    getYamlContent,
+    isCfInstalled,
+    isLoggedInCf,
+    isMtaProject,
     loadApps,
-    type AttributesAnswers,
-    type ConfigAnswers,
-    type UI5Version
+    loadCfConfig
 } from '@sap-ux/adp-tooling';
-import type { AbapServiceProvider } from '@sap-ux/axios-extension';
-import { isInternalFeaturesSettingEnabled } from '@sap-ux/feature-toggle';
 import {
     TelemetryHelper,
     getDefaultTargetFolder,
@@ -27,6 +29,9 @@ import {
 } from '@sap-ux/fiori-generator-shared';
 import { ToolsLogger } from '@sap-ux/logger';
 import type { Manifest } from '@sap-ux/project-access';
+import type { AbapServiceProvider } from '@sap-ux/axios-extension';
+import { isInternalFeaturesSettingEnabled } from '@sap-ux/feature-toggle';
+import type { CfConfig, CfServicesAnswers, AttributesAnswers, ConfigAnswers, UI5Version } from '@sap-ux/adp-tooling';
 
 import { EventName } from '../telemetryEvents';
 import { cacheClear, cacheGet, cachePut, initCache } from '../utils/appWizardCache';
@@ -35,16 +40,31 @@ import { initI18n, t } from '../utils/i18n';
 import AdpGeneratorLogger from '../utils/logger';
 import { setHeaderTitle } from '../utils/opts';
 import { getFirstArgAsString, parseJsonInput } from '../utils/parse-json-input';
-import { getDeployPage, getWizardPages, updateFlpWizardSteps, updateWizardSteps } from '../utils/steps';
+import {
+    getDeployPage,
+    getWizardPages,
+    updateCfWizardSteps,
+    updateFlpWizardSteps,
+    updateWizardSteps
+} from '../utils/steps';
 import { addDeployGen, addExtProjectGen, addFlpGen } from '../utils/subgenHelpers';
 import { getTemplatesOverwritePath } from '../utils/templates';
 import { existsInWorkspace, handleWorkspaceFolderChoice, showWorkspaceFolderWarning } from '../utils/workspace';
 import { getFlexLayer } from './layer';
 import { getPrompts } from './questions/attributes';
+import { CFServicesPrompter } from './questions/cf-services';
 import { ConfigPrompter } from './questions/configuration';
 import { getDefaultNamespace, getDefaultProjectName } from './questions/helper/default-values';
 import { validateJsonInput } from './questions/helper/validators';
-import { type AdpGeneratorOptions, type AttributePromptOptions, type JsonInput } from './types';
+import {
+    TargetEnv,
+    type TargetEnvAnswers,
+    type AdpGeneratorOptions,
+    type AttributePromptOptions,
+    type JsonInput
+} from './types';
+import { isAppStudio } from '@sap-ux/btp-utils';
+import { getProjectPathPrompt, getTargetEnvPrompt } from './questions/target-env';
 
 const generatorTitle = 'Adaptation Project';
 
@@ -97,6 +117,10 @@ export default class extends Generator {
      */
     private prompter: ConfigPrompter;
     /**
+     * Instance of the CF services prompter class.
+     */
+    private cfPrompter: CFServicesPrompter;
+    /**
      * JSON object representing the complete adaptation project configuration,
      * passed as a CLI argument.
      */
@@ -117,6 +141,38 @@ export default class extends Generator {
      * Indicates if the current layer is based on a customer base.
      */
     private isCustomerBase: boolean;
+    /**
+     * Target environment.
+     */
+    private targetEnv: TargetEnv;
+    /**
+     * Indicates if the current environment is a CF environment.
+     */
+    private isCfEnv = false;
+    /**
+     * Indicates if the user is logged in to CF.
+     */
+    private isCfLoggedIn = false;
+    /**
+     * CF config.
+     */
+    private cfConfig: CfConfig;
+    /**
+     * Indicates if the current project is an MTA project.
+     */
+    private readonly isMtaYamlFound: boolean;
+    /**
+     * CF services answers.
+     */
+    private cfServicesAnswers: CfServicesAnswers;
+    /**
+     * Indicates if the extension is installed.
+     */
+    private readonly isExtensionInstalled: boolean;
+    /**
+     * Indicates if CF is installed.
+     */
+    private cfInstalled: boolean;
 
     /**
      * Creates an instance of the generator.
@@ -130,9 +186,13 @@ export default class extends Generator {
         this.shouldInstallDeps = opts.shouldInstallDeps ?? true;
         this.toolsLogger = new ToolsLogger();
         this.vscode = opts.vscode;
+        this._setupLogging();
         this.options = opts;
 
-        this._setupLogging();
+        this.isMtaYamlFound = isMtaProject(process.cwd()) as boolean;
+        this.isExtensionInstalled = isInternalFeaturesSettingEnabled()
+            ? isExtensionInstalled(opts.vscode, 'SAP.adp-ve-bas-ext')
+            : false;
 
         const jsonInputString = getFirstArgAsString(args);
         this.jsonInput = parseJsonInput(jsonInputString, this.logger);
@@ -161,9 +221,17 @@ export default class extends Generator {
         this.isCustomerBase = this.layer === FlexLayer.CUSTOMER_BASE;
         this.systemLookup = new SystemLookup(this.logger);
 
+        this.cfInstalled = await isCfInstalled(this.logger);
+        this.cfConfig = loadCfConfig(this.logger);
+        this.isCfLoggedIn = await isLoggedInCf(this.cfConfig, this.logger);
+        this.logger.info(`isCfInstalled: ${this.cfInstalled}`);
+
+        const isInternalUsage = isInternalFeaturesSettingEnabled();
         if (!this.jsonInput) {
-            this.prompts.splice(0, 0, getWizardPages());
+            const shouldShowTargetEnv = isAppStudio() && this.cfInstalled && this.isExtensionInstalled;
+            this.prompts.splice(0, 0, getWizardPages(shouldShowTargetEnv));
             this.prompter = this._getOrCreatePrompter();
+            this.cfPrompter = new CFServicesPrompter(isInternalUsage, this.isCfLoggedIn, this.logger);
         }
 
         await TelemetryHelper.initTelemetrySettings({
@@ -171,7 +239,7 @@ export default class extends Generator {
                 name: '@sap/generator-fiori:generator-adp',
                 version: this.rootGeneratorVersion()
             },
-            internalFeature: isInternalFeaturesSettingEnabled(),
+            internalFeature: isInternalUsage,
             watchTelemetrySettingStore: false
         });
     }
@@ -181,75 +249,87 @@ export default class extends Generator {
             return;
         }
 
-        const isExtensibilityExtInstalled = isExtensionInstalled(this.vscode, 'SAP.vscode-bas-extensibility');
-        const configQuestions = this.prompter.getPrompts({
-            appValidationCli: { hide: !this.isCli },
-            systemValidationCli: { hide: !this.isCli },
-            shouldCreateExtProject: { isExtensibilityExtInstalled }
-        });
-        this.configAnswers = await this.prompt<ConfigAnswers>(configQuestions);
-        this.shouldCreateExtProject = !!this.configAnswers.shouldCreateExtProject;
+        await this._determineTargetEnv();
 
-        this.logger.info(`System: ${this.configAnswers.system}`);
-        this.logger.info(`Application: ${JSON.stringify(this.configAnswers.application, null, 2)}`);
+        if (this.isCfEnv) {
+            await this._promptForCfEnvironment();
+        } else {
+            const isExtensibilityExtInstalled = isExtensionInstalled(this.vscode, 'SAP.vscode-bas-extensibility');
+            const configQuestions = this.prompter.getPrompts({
+                appValidationCli: { hide: !this.isCli },
+                systemValidationCli: { hide: !this.isCli },
+                shouldCreateExtProject: { isExtensibilityExtInstalled }
+            });
+            this.configAnswers = await this.prompt<ConfigAnswers>(configQuestions);
+            this.shouldCreateExtProject = !!this.configAnswers.shouldCreateExtProject;
 
-        const { ui5Versions, systemVersion } = this.prompter.ui5;
-        const promptConfig = {
-            ui5Versions,
-            isVersionDetected: !!systemVersion,
-            isCloudProject: this.prompter.isCloud,
-            layer: this.layer,
-            prompts: this.prompts
-        };
-        const defaultFolder = getDefaultTargetFolder(this.options.vscode) ?? process.cwd();
-        const options: AttributePromptOptions = {
-            targetFolder: { default: defaultFolder, hide: this.shouldCreateExtProject },
-            ui5ValidationCli: { hide: !this.isCli },
-            enableTypeScript: { hide: this.shouldCreateExtProject },
-            addFlpConfig: { hasBaseAppInbounds: !!this.prompter.baseAppInbounds, hide: this.shouldCreateExtProject },
-            addDeployConfig: { hide: this.shouldCreateExtProject || !this.isCustomerBase }
-        };
-        const attributesQuestions = getPrompts(this.destinationPath(), promptConfig, options);
+            this.logger.info(`System: ${this.configAnswers.system}`);
+            this.logger.info(`Application: ${JSON.stringify(this.configAnswers.application, null, 2)}`);
 
-        this.attributeAnswers = await this.prompt(attributesQuestions);
-
-        // Steps need to be updated here to be available after back navigation in Yeoman UI.
-        this._updateWizardStepsAfterNavigation();
-
-        this.logger.info(`Project Attributes: ${JSON.stringify(this.attributeAnswers, null, 2)}`);
-
-        if (this.attributeAnswers.addDeployConfig) {
-            const system = await this.systemLookup.getSystemByName(this.configAnswers.system);
-            addDeployGen(
-                {
-                    projectName: this.attributeAnswers.projectName,
-                    projectPath: this.attributeAnswers.targetFolder,
-                    connectedSystem: this.configAnswers.system,
-                    system
+            const { ui5Versions, systemVersion } = this.prompter.ui5;
+            const promptConfig = {
+                ui5Versions,
+                isVersionDetected: !!systemVersion,
+                isCloudProject: this.prompter.isCloud,
+                layer: this.layer,
+                prompts: this.prompts
+            };
+            const defaultFolder = getDefaultTargetFolder(this.options.vscode) ?? process.cwd();
+            const options: AttributePromptOptions = {
+                targetFolder: { default: defaultFolder, hide: this.shouldCreateExtProject },
+                ui5ValidationCli: { hide: !this.isCli },
+                enableTypeScript: { hide: this.shouldCreateExtProject },
+                addFlpConfig: {
+                    hasBaseAppInbounds: !!this.prompter.baseAppInbounds,
+                    hide: this.shouldCreateExtProject
                 },
-                this.composeWith.bind(this),
-                this.logger,
-                this.appWizard
-            );
-        }
+                addDeployConfig: { hide: this.shouldCreateExtProject || !this.isCustomerBase }
+            };
+            const attributesQuestions = getPrompts(this.destinationPath(), promptConfig, options);
+            this.attributeAnswers = await this.prompt(attributesQuestions);
 
-        if (this.attributeAnswers?.addFlpConfig) {
-            addFlpGen(
-                {
-                    vscode: this.vscode,
-                    projectRootPath: this._getProjectPath(),
-                    inbounds: this.prompter.baseAppInbounds,
-                    layer: this.layer
-                },
-                this.composeWith.bind(this),
-                this.logger,
-                this.appWizard
-            );
+            // Steps need to be updated here to be available after back navigation in Yeoman UI.
+            this._updateWizardStepsAfterNavigation();
+
+            this.logger.info(`Project Attributes: ${JSON.stringify(this.attributeAnswers, null, 2)}`);
+            if (this.attributeAnswers.addDeployConfig) {
+                const system = await this.systemLookup.getSystemByName(this.configAnswers.system);
+                addDeployGen(
+                    {
+                        projectName: this.attributeAnswers.projectName,
+                        projectPath: this.attributeAnswers.targetFolder,
+                        connectedSystem: this.configAnswers.system,
+                        system
+                    },
+                    this.composeWith.bind(this),
+                    this.logger,
+                    this.appWizard
+                );
+            }
+
+            if (this.attributeAnswers?.addFlpConfig) {
+                addFlpGen(
+                    {
+                        vscode: this.vscode,
+                        projectRootPath: this._getProjectPath(),
+                        inbounds: this.prompter.baseAppInbounds,
+                        layer: this.layer
+                    },
+                    this.composeWith.bind(this),
+                    this.logger,
+                    this.appWizard
+                );
+            }
         }
     }
 
     async writing(): Promise<void> {
         try {
+            if (this.isCfEnv) {
+                await this._generateAdpProjectArtifactsCF();
+                return;
+            }
+
             if (this.jsonInput) {
                 await this._initFromJson();
             }
@@ -299,7 +379,7 @@ export default class extends Generator {
     }
 
     async install(): Promise<void> {
-        if (!this.shouldInstallDeps || this.shouldCreateExtProject) {
+        if (!this.shouldInstallDeps || this.shouldCreateExtProject || this.isCfEnv) {
             return;
         }
 
@@ -311,7 +391,7 @@ export default class extends Generator {
     }
 
     async end(): Promise<void> {
-        if (this.shouldCreateExtProject) {
+        if (this.shouldCreateExtProject || this.isCfEnv) {
             return;
         }
 
@@ -327,7 +407,7 @@ export default class extends Generator {
             });
         }
 
-        if (isCFEnvironment(projectPath) || this.isCli) {
+        if (this.isCli) {
             return;
         }
 
@@ -347,6 +427,91 @@ export default class extends Generator {
     }
 
     /**
+     * Prompts the user for the CF project path.
+     */
+    private async _promptForCfProjectPath(): Promise<void> {
+        if (this.isMtaYamlFound) {
+            const path = this.destinationRoot(process.cwd());
+            getYamlContent(join(path, 'mta.yaml'));
+            this.logger.log(`Project path information: ${path}`);
+        } else {
+            const pathAnswers = await this.prompt([getProjectPathPrompt(this.logger, this.vscode)]);
+            const path = this.destinationRoot(fs.realpathSync(pathAnswers.projectLocation, 'utf-8'));
+            this.logger.log(`Project path information: ${path}`);
+        }
+    }
+
+    /**
+     * Determines the target environment based on the current context.
+     * Sets the target environment and updates related state accordingly.
+     */
+    private async _determineTargetEnv(): Promise<void> {
+        const hasRequiredExtensions = this.isExtensionInstalled && this.cfInstalled;
+
+        if (isAppStudio() && hasRequiredExtensions) {
+            await this._promptForTargetEnvironment();
+        } else {
+            this.targetEnv = TargetEnv.ABAP;
+        }
+    }
+
+    /**
+     * Prompts the user to select the target environment and updates related state.
+     */
+    private async _promptForTargetEnvironment(): Promise<void> {
+        const targetEnvAnswers = await this.prompt<TargetEnvAnswers>([
+            getTargetEnvPrompt(this.appWizard, this.cfInstalled, this.isCfLoggedIn, this.cfConfig, this.vscode)
+        ]);
+
+        this.targetEnv = targetEnvAnswers.targetEnv;
+        this.isCfEnv = this.targetEnv === TargetEnv.CF;
+        this.logger.info(`Target environment: ${this.targetEnv}`);
+
+        updateCfWizardSteps(this.isCfEnv, this.prompts);
+
+        this.logger.log(`Project organization information: ${JSON.stringify(this.cfConfig.org, null, 2)}`);
+        this.logger.log(`Project space information: ${JSON.stringify(this.cfConfig.space, null, 2)}`);
+        this.logger.log(`Project apiUrl information: ${JSON.stringify(this.cfConfig.url, null, 2)}`);
+    }
+
+    /**
+     * Prompts the user for the CF environment.
+     */
+    private async _promptForCfEnvironment(): Promise<void> {
+        await this._promptForCfProjectPath();
+
+        const options: AttributePromptOptions = {
+            targetFolder: { hide: true },
+            ui5Version: { hide: true },
+            ui5ValidationCli: { hide: true },
+            enableTypeScript: { hide: true },
+            addFlpConfig: { hide: true },
+            addDeployConfig: { hide: true }
+        };
+
+        const projectPath = this.destinationPath();
+        const attributesQuestions = getPrompts(
+            projectPath,
+            {
+                ui5Versions: [],
+                isVersionDetected: false,
+                isCloudProject: false,
+                layer: this.layer,
+                prompts: this.prompts,
+                isCfEnv: true
+            },
+            options
+        );
+
+        this.attributeAnswers = await this.prompt(attributesQuestions);
+        this.logger.info(`Project Attributes: ${JSON.stringify(this.attributeAnswers, null, 2)}`);
+
+        const cfServicesQuestions = await this.cfPrompter.getPrompts(projectPath, this.cfConfig);
+        this.cfServicesAnswers = await this.prompt<CfServicesAnswers>(cfServicesQuestions);
+        this.logger.info(`CF Services Answers: ${JSON.stringify(this.cfServicesAnswers, null, 2)}`);
+    }
+
+    /**
      * Retrieves the ConfigPrompter instance from cache if it exists, otherwise creates a new instance.
      *
      * @returns {ConfigPrompter} Cached config prompter if going back a page.
@@ -360,6 +525,37 @@ export default class extends Generator {
         const prompter = new ConfigPrompter(this.systemLookup, this.layer, this.logger);
         cachePut(this.appWizard, { prompter }, this.logger);
         return prompter;
+    }
+
+    /**
+     * Generates the ADP project artifacts for the CF environment.
+     */
+    private async _generateAdpProjectArtifactsCF(): Promise<void> {
+        const projectPath = this.isMtaYamlFound ? process.cwd() : this.destinationPath();
+        const publicVersions = await fetchPublicVersions(this.logger);
+
+        const manifest = this.cfPrompter.manifest;
+        if (!manifest) {
+            throw new Error('Manifest not found for base app.');
+        }
+
+        const html5RepoRuntimeGuid = this.cfPrompter.serviceInstanceGuid;
+        const config = getCfConfig({
+            attributeAnswers: this.attributeAnswers,
+            cfServicesAnswers: this.cfServicesAnswers,
+            cfConfig: this.cfConfig,
+            layer: this.layer,
+            manifest,
+            html5RepoRuntimeGuid,
+            projectPath,
+            publicVersions
+        });
+
+        if (config.options) {
+            config.options.templatePathOverwrite = getTemplatesOverwritePath();
+        }
+
+        await generateCf(projectPath, config, this.logger, this.fs);
     }
 
     /**