+++ packages/app-config-writer/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @sap-ux/app-config-writer
 
+## 0.6.45
+
+### Patch Changes
+
+-   Updated dependencies [bacaf93]
+    -   @sap-ux/axios-extension@1.23.0
+    -   @sap-ux/store@1.2.0
+    -   @sap-ux/ui5-config@0.29.8
+
+## 0.6.44
+
+### Patch Changes
+
+-   99155bf: fix: configuration source message
+
+## 0.6.43
+
+### Patch Changes
+
+-   Updated dependencies [9e94382]
+    -   @sap-ux/ui5-config@0.29.8
+    -   @sap-ux/project-access@1.32.4
+    -   @sap-ux/axios-extension@1.22.10
+
 ## 0.6.42
 
 ### Patch Changes