+++ packages/environment-check/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux/environment-check
 
+## 0.18.37
+
+### Patch Changes
+
+-   Updated dependencies [bacaf93]
+    -   @sap-ux/axios-extension@1.23.0
+    -   @sap-ux/store@1.2.0
+    -   @sap-ux/ui5-config@0.29.8
+
+## 0.18.36
+
+### Patch Changes
+
+-   Updated dependencies [9e94382]
+    -   @sap-ux/ui5-config@0.29.8
+    -   @sap-ux/project-access@1.32.4
+    -   @sap-ux/axios-extension@1.22.10
+
 ## 0.18.35
 
 ### Patch Changes