+++ packages/ui-components/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/ui-components
 
+## 1.26.13
+
+### Patch Changes
+
+-   6debb89: style update for disabled combobox option
+
+## 1.26.12
+
+### Patch Changes
+
+-   4c448a7: UIDialog: updated interface to support multiple elements within the footer
+
 ## 1.26.11
 
 ### Patch Changes